Take mharju's suggestion to explicitly use int64_t

+++ src/parquet_cursor.cc
@@ -752,7 +752,7 @@ void ParquetCursor::ensureColumn(int col) {
       }
       case parquet::Type::INT64: {
         parquet::Int64Scanner *s = (parquet::Int64Scanner *)scanners[col].get();
-        long long rv = 0;
+        int64_t rv = 0;
         s->NextValue(&rv, &wasNull);
         break;
       }
@@ -830,7 +830,7 @@ void ParquetCursor::ensureColumn(int col) {
     }
     case parquet::Type::INT64: {
       parquet::Int64Scanner *s = (parquet::Int64Scanner *)scanners[col].get();
-      long long rv = 0;
+      int64_t rv = 0;
       hadValue = s->NextValue(&rv, &wasNull);
       colIntValues[col] = rv;
       break;
